chore: renew `New Line Before Return` rule
