+++ lib/src/rules/newline_before_return/return_statement_visitor.dart
@@ -1,0 +1,13 @@
+part of 'newline_before_return.dart';
+
+class _Visitor extends RecursiveAstVisitor<void> {
+  final _statements = <ReturnStatement>[];
+
+  Iterable<ReturnStatement> get statements => _statements;
+
+  @override
+  void visitReturnStatement(ReturnStatement node) {
+    super.visitReturnStatement(node);
+    _statements.add(node);
+  }
+}
